moves init apex from LM to apex connector (#3923)
